Basic homebrew formula for devlog v0.2.0
Add basic homebrew formula for devlog v0.2.0.
This downloads the release tarball, installs the binary,
and prints instructions for adding vim syntax plugins
to the vim runtimepath.

The formula doesn't have a test block because currently
it's hard to test the devlog tool without writing to stdin.
I'm planning to add flags to the devlog tool for non-interactive
use, which will make writing the tests easier.